# docs/server-api.openapi.yaml
openapi: 3.0.3
info:
  title: ManyRows Server-to-Server API
  version: "1.0.0"
  description: >
    The customer backend (server-to-server) API. Authenticate every request
    with a workspace API key in the `X-API-Key` header (or
    `Authorization: Bearer <key>`). Every endpoint is scoped to **members of the
    calling app**: the user pool only shares credentials across apps (SSO), it is
    not an access boundary, so a key for one app only sees and acts on users who
    have joined that app. For isolation between apps, give them separate pools.

    Requests are rate limited per API key (HTTP 429 + `Retry-After` when
    exceeded). Errors are returned as JSON `{ "error": "<code>", "message": "..." }`.
servers:
  - url: https://{host}/x/{workspaceSlug}/api/v1
    description: A workspace's server API base.
    variables:
      host:
        default: auth.example.com
        description: Your ManyRows host.
      workspaceSlug:
        default: your-workspace
        description: The workspace slug.
security:
  - ApiKeyAuth: []
...
paths:
  /apps/{appId}:
    parameters:
      - $ref: "#/components/parameters/AppId"
    get:
      tags: [Delivery]
      summary: Runtime delivery
      description: >
        All config values and feature flags for the app. Supports `ETag` /
        `If-None-Match` (returns `304 Not Modified` when unchanged).
      operationId: getDelivery
      responses:
        "200":
          description: Config and flags.
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeliveryResponse" }
        "304": { description: Not modified (matched `If-None-Match`). }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /apps/{appId}/check-permission:
    parameters:
      - $ref: "#/components/parameters/AppId"
    get:
      tags: [Authorization]
      summary: Check a user's permission
      operationId: checkPermission
      parameters:
        - name: accountId
          in: query
          required: true
          description: The user ID (a member of this app) to check.
          schema: { type: string, format: uuid }
        - name: permission
          in: query
          required: true
          description: Permission slug, e.g. `posts:read`.
          schema: { type: string }
      responses:
        "200":
          description: Whether the user has the permission.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckPermissionResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /apps/{appId}/roles:
    parameters:
      - $ref: "#/components/parameters/AppId"
    get:
      tags: [Authorization]
      summary: List roles
      description: The product's roles, each with the permission slugs it grants.
      operationId: listRoles
      responses:
        "200":
          description: Role catalog.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RolesListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /apps/{appId}/permissions:
    parameters:
      - $ref: "#/components/parameters/AppId"
    get:
      tags: [Authorization]
      summary: List permissions
      operationId: listPermissions
      responses:
        "200":
          description: Permission catalog.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PermissionsListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /apps/{appId}/users:
    parameters:
      - $ref: "#/components/parameters/AppId"
    get:
      tags: [Users]
      summary: List members or look one up by email
      description: >
        With no `email`, lists the app's members (paginated, `search` substring
        filter) and returns a `MembersListResponse`. With `email=<addr>`, returns
        that single member as a `ServerUserResponse`.
      operationId: listOrLookupUsers
      parameters:
        - name: email
          in: query
          description: Exact email — returns a single deep user instead of a list.
          schema: { type: string, format: email }
        - name: search
          in: query
          description: Email substring filter for the list.
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: A member list, or a single user when `email` is given.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/MembersListResponse"
                  - $ref: "#/components/schemas/ServerUserResponse"
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Users]
      summary: Provision a user
      description: >
        Create-or-find a user by email in the app's pool and add them to the app,
        in one idempotent call. Returns `201` for a new identity, `200` when an
        existing one was reused.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: A new identity was created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateUserResponse" }
        "200":
          description: An existing identity was reused (now a member).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateUserResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /apps/{appId}/users:batch:
    parameters:
      - $ref: "#/components/parameters/AppId"
    post:
      tags: [Users]
      summary: Provision many users at once
      description: >
        Provision up to 100 users in one call, all with the same optional role
        set. Roles are resolved once (a bad slug fails the whole request); each
        email is then provisioned independently and reported in `results`, so
        one bad email doesn't sink the rest. Idempotent per email.
      operationId: batchCreateUsers
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BatchCreateUsersRequest" }
      responses:
        "200":
          description: Per-email results.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BatchCreateUsersResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /apps/{appId}/users/{userId}:
    parameters:
      - $ref: "#/components/parameters/AppId"
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [Users]
      summary: Get a user by ID
      operationId: getUser
      responses:
        "200":
          description: The user with roles, permissions, and field values.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerUserResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Users]
      summary: Suspend or re-enable a user
      description: >
        Set the user's status in this app. `disabled` blocks sign-in to this app
        and revokes its active sessions; `active` restores access. Per-app — a
        user shared across apps in one pool is unaffected elsewhere.
      operationId: setUserStatus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetUserStatusRequest" }
      responses:
        "200":
          description: Updated status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserStatusResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Users]
      summary: Remove a user from the app
      description: >
        Drops the user's membership, roles, permission overrides, and sessions
        for this app. If that leaves the user in no other app, the pool identity
        is deleted too; otherwise it's kept.
      operationId: removeUser
      responses:
        "200":
          description: Removed; says whether the identity was also deleted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RemoveUserResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /apps/{appId}/users/{userId}/roles:
    parameters:
      - $ref: "#/components/parameters/AppId"
      - $ref: "#/components/parameters/UserId"
    put:
      tags: [Users]
      summary: Replace a user's roles
      description: >
        Full replace (not merge), by role slug. An empty array clears all roles
        and revokes the user's sessions for the app.
      operationId: replaceUserRoles
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReplaceRolesRequest" }
      responses:
        "200":
          description: The resulting role slugs.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RolesResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /apps/{appId}/users/{userId}/permissions:
    parameters:
      - $ref: "#/components/parameters/AppId"
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [Users]
      summary: List a user's direct permission overrides
      description: Per-user grants, separate from the permissions inherited via roles.
      operationId: getUserPermissions
      responses:
        "200":
          description: Direct permission slugs.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPermissionsResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Users]
      summary: Replace a user's direct permission overrides
      operationId: setUserPermissions
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetPermissionsRequest" }
      responses:
        "200":
          description: The resulting permission slugs.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPermissionsResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /apps/{appId}/users/{userId}/sessions:
    parameters:
      - $ref: "#/components/parameters/AppId"
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [Users]
      summary: List a user's sessions for this app
      operationId: listUserSessions
      responses:
        "200":
          description: The user's active sessions in this app.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionsListResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Users]
      summary: Revoke all a user's sessions (force-logout)
      operationId: revokeUserSessions
      responses:
        "200":
          description: Count of sessions revoked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RevokeSessionsResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    AppId:
      name: appId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    DeliveryResponse:
      type: object
      properties:
        workspaceId: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        appId: { type: string, format: uuid }
        updatedAt: { type: string, format: date-time }
        config: { $ref: "#/components/schemas/DeliveryConfigBuckets" }
        flags: { $ref: "#/components/schemas/DeliveryFlagBuckets" }
...